# backend/app/swagger.yaml
swagger: "2.0"
info:
  title: School Management System API
  description: |
    REST API for the School Management System (SMS).

    ## Authentication
    Most endpoints require a JWT Bearer token. Log in via **POST /auth/login** to obtain
    `access_token` and `refresh_token`. Pass the access token as:
    ```
    Authorization: Bearer <access_token>
    ```
    The access token expires in **15 minutes**. Use **POST /auth/refresh** with the refresh token
    to get a new access token (refresh tokens expire in 7 days).

    ## Roles
    | Role | Access |
    |------|--------|
    | `admin` | All endpoints |
    | `teacher` | Students, attendance, grades, leave review |
    | `student` | Own data only |
    | `parent` | Parent portal, leave submissions, notifications |

    ## Response Envelope
    All responses follow:
    ```json
    { "success": true, "data": {}, "message": "...", "errors": null }
    ```

    ## Dev Seed Credentials
    | Role | Email | Password |
    |------|-------|----------|
    | Admin | admin@school.sms | Admin@1234 |
    | Teacher | priya.sharma@school.sms | Teacher@123 |
    | Student | alice.johnson@school.sms | Student@123 |
    | Parent | robert.johnson@parents.school.sms | Parent@123 |

  version: "1.0.0"
  contact:
    email: admin@school.sms
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Format: Bearer <JWT access token>"
...
definitions:
...
  User:
    type: object
    properties:
      id:
        type: integer
        example: 1
      email:
        type: string
        example: admin@school.sms
      role:
        type: string
        enum: [admin, teacher, student, parent]
        example: admin
      first_name:
        type: string
        example: Rahul
      last_name:
        type: string
        example: Verma
      full_name:
        type: string
        example: Rahul Verma
      photo_url:
        type: string
        nullable: true
        example: /uploads/profile_photos/user_1_abc123.jpg
      is_active:
        type: boolean
        example: true
      last_login:
        type: string
        format: date-time
        nullable: true
      created_at:
        type: string
        format: date-time
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      description: Paginated list of users with optional role/search filter. **Admin only.**
      security:
        - Bearer: []
      parameters:
        - in: query
          name: page
          type: integer
          default: 1
        - in: query
          name: per_page
          type: integer
          default: 20
          maximum: 100
        - in: query
          name: role
          type: string
          enum: [admin, teacher, student, parent]
          description: Filter by role
        - in: query
          name: q
          type: string
          description: Search by email, first_name, or last_name
        - in: query
          name: is_active
          type: boolean
          description: Filter active/inactive users
      responses:
        200:
          description: List of users
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/definitions/User'
        401:
          description: Unauthorized
        403:
          description: Admin access required

    post:
      tags: [Users]
      summary: Create a new user
      description: |
        Create a user account. **Admin only.**

        - For `role=parent`: also requires `relationship_type` and `phone_primary` — a Parent profile is auto-created.
        - For `role=teacher` or `role=student`: only the User record is created (full profiles via their own endpoints).

        **Password requirements:** 8+ chars, uppercase, lowercase, digit, special character.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [email, password, role, first_name, last_name]
            properties:
              email:
                type: string
                example: teacher1@school.sms
              password:
                type: string
                example: Teacher@123
              role:
                type: string
                enum: [admin, teacher, student, parent]
                example: teacher
              first_name:
                type: string
                example: Priya
              last_name:
                type: string
                example: Sharma
              relationship_type:
                type: string
                enum: [Father, Mother, Guardian]
                description: Required if role=parent
              phone_primary:
                type: string
                description: Required if role=parent
                example: "+91-9988776655"
              phone_secondary:
                type: string
              occupation:
                type: string
              address:
                type: string
      responses:
        201:
          description: User created
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  user:
                    $ref: '#/definitions/User'
        400:
          description: Missing required fields or invalid role
        401:
          description: Unauthorized
        403:
          description: Admin access required
        409:
          description: Email already exists
        422:
          description: Weak password or invalid email format
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Retrieve a single user by ID. **Admin only.**
      security:
        - Bearer: []
      parameters:
        - in: path
          name: user_id
          type: integer
          required: true
          example: 1
      responses:
        200:
          description: User found
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/User'
        401:
          description: Unauthorized
        403:
          description: Admin access required
        404:
          description: User not found

    delete:
      tags: [Users]
      summary: Deactivate a user (soft delete)
      description: |
        Sets `is_active=false` on the user. Does not delete the record.
        **Admin only.** Cannot deactivate your own account.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: user_id
          type: integer
          required: true
          example: 3
      responses:
        200:
          description: User deactivated
        400:
          description: Cannot deactivate own account or already inactive
        401:
          description: Unauthorized
        403:
          description: Admin access required
        404:
          description: User not found

  # ── Students ─────────────────────────────────────────────────────────────────
...